# internal/platform/docs/swagger.yaml
basePath: /api
definitions:
  httptransport.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v1.ChangePasswordRequest:
    properties:
      confirm_password:
        type: string
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - confirm_password
    - current_password
    - new_password
    type: object
...
  v1.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  v1.SessionListResponse:
    properties:
      pagination:
        $ref: '#/definitions/v1.Pagination'
      sessions:
        items:
          $ref: '#/definitions/v1.SessionSession'
        type: array
    type: object
  v1.SessionSession:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      last_seen:
        type: string
      token:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
    type: object
...
  v1.UpdateProfileRequest:
    properties:
      avatar:
        type: string
      bio:
        type: string
      email:
        type: string
      phone:
        type: string
      settings:
        $ref: '#/definitions/v1.UserSettings'
    type: object
  v1.UserInfo:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        description: active, inactive, suspended
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  v1.UserListResponse:
    properties:
      pagination:
        $ref: '#/definitions/v1.Pagination'
      users:
        items:
          $ref: '#/definitions/v1.UserProfile'
        type: array
    type: object
  v1.UserProfile:
    properties:
      avatar:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      phone:
        type: string
      settings:
        $ref: '#/definitions/v1.UserSettings'
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  v1.UserSettings:
    properties:
      language:
        type: string
      notifications:
        type: boolean
      theme:
        type: string
      timezone:
        type: string
    type: object
...
  webapi.AuthResponse:
    properties:
      expires_at:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/webapi.UserInfo'
    type: object
...
  webapi.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  webapi.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  webapi.UserInfo:
    properties:
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 小智服务端，包含OTA与Vision等接口
  title: 小智服务端 API 文档
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/webapi.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webapi.AuthResponse'
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
      summary: User login
      tags:
      - Auth
...
  /auth/refresh:
    post:
      description: Refresh the authentication token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webapi.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Refresh token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/webapi.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webapi.AuthResponse'
        "400":
          description: Bad Request
          schema:
            type: object
        "409":
          description: Conflict
          schema:
            type: object
      summary: User registration
      tags:
      - Auth
...
  /sessions:
    get:
      description: 获取当前用户的会话列表
      parameters:
      - description: 用户ID
        format: int64
        in: query
        name: user_id
        type: integer
      - description: 是否只返回活跃会话
        in: query
        name: active
        type: boolean
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.SessionListResponse'
              type: object
      summary: 获取会话列表
      tags:
      - Sessions
  /sessions/{id}:
    delete:
      description: 撤销指定的用户会话
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 撤销会话
      tags:
      - Sessions
...
  /users:
    get:
      description: 获取用户列表，支持分页和过滤
      parameters:
      - description: 按角色过滤
        in: query
        name: role
        type: string
      - description: 按状态过滤
        in: query
        name: status
        type: string
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: created_at
        description: 排序字段
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序方向
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.UserListResponse'
              type: object
      summary: 获取用户列表
      tags:
      - Users
  /users/{id}:
    get:
      description: 根据ID获取特定用户的详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.UserInfo'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 获取用户信息
      tags:
      - Users
  /users/change-password:
    post:
      consumes:
      - application/json
      description: 修改当前登录用户的密码
      parameters:
      - description: 密码修改信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 修改密码
      tags:
      - Users
  /users/profile:
    get:
      description: 获取当前登录用户的详细档案信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.UserProfile'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 获取用户档案
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: 更新当前登录用户的档案信息
      parameters:
      - description: 档案更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.UserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 更新用户档案
      tags:
      - Users
...
swagger: "2.0"